Refactor gravity simulation into package
