+++ src/tanks/gravity_sim/__init__.py
@@ -1,0 +1,17 @@
+"""Gravity simulation package."""
+
+from .app import run
+from .feedback import DensityFeedbackController
+from .point import GravityPoint
+from .settings import SimulationSettings
+from .simulation import GravitySimulation
+from .slider import Slider
+
+__all__ = [
+    "run",
+    "DensityFeedbackController",
+    "GravityPoint",
+    "SimulationSettings",
+    "GravitySimulation",
+    "Slider",
+]
