+++ src/tanks/gravity_sim/constants.py
@@ -1,0 +1,18 @@
+"""Shared constants for the gravity simulation UI."""
+
+SCREEN_WIDTH = 2200
+SCREEN_HEIGHT = 1240
+CONTROL_PANEL_WIDTH = 260
+SIMULATION_WIDTH = SCREEN_WIDTH - CONTROL_PANEL_WIDTH
+BACKGROUND_COLOR = (15, 15, 30)
+PANEL_COLOR = (28, 28, 48)
+TEXT_COLOR = (220, 220, 230)
+SLIDER_TRACK_COLOR = (80, 80, 110)
+SLIDER_HANDLE_COLOR = (210, 210, 230)
+TRAIL_MAX_LENGTH = 1
+
+GROUP_COLORS: list[tuple[int, int, int]] = [
+    (236, 120, 120),
+    (118, 210, 180),
+    (130, 156, 246),
+]
